// client/ts/generated/types/messageAddressTableLookup.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  addDecoderSizePrefix,
  addEncoderSizePrefix,
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getBytesDecoder,
  getBytesEncoder,
  getStructDecoder,
  getStructEncoder,
  getU32Decoder,
  getU32Encoder,
  type Address,
  type Codec,
  type Decoder,
  type Encoder,
  type ReadonlyUint8Array,
} from '@solana/kit';

/**
 * Address table lookups describe an on-chain address lookup table to use
 * for loading more readonly and writable accounts into a transaction.
 */
export type MessageAddressTableLookup = {
  /** Address lookup table account key. */
  accountKey: Address;
  /** List of indexes used to load writable accounts. */
  writableIndexes: ReadonlyUint8Array;
  /** List of indexes used to load readonly accounts. */
  readonlyIndexes: ReadonlyUint8Array;
};

export type MessageAddressTableLookupArgs = MessageAddressTableLookup;

export function getMessageAddressTableLookupEncoder(): Encoder<MessageAddressTableLookupArgs> {
  return getStructEncoder([
    ['accountKey', getAddressEncoder()],
    [
      'writableIndexes',
      addEncoderSizePrefix(getBytesEncoder(), getU32Encoder()),
    ],
    [
      'readonlyIndexes',
      addEncoderSizePrefix(getBytesEncoder(), getU32Encoder()),
    ],
  ]);
}

export function getMessageAddressTableLookupDecoder(): Decoder<MessageAddressTableLookup> {
  return getStructDecoder([
    ['accountKey', getAddressDecoder()],
    [
      'writableIndexes',
      addDecoderSizePrefix(getBytesDecoder(), getU32Decoder()),
    ],
    [
      'readonlyIndexes',
      addDecoderSizePrefix(getBytesDecoder(), getU32Decoder()),
    ],
  ]);
}

export function getMessageAddressTableLookupCodec(): Codec<
  MessageAddressTableLookupArgs,
  MessageAddressTableLookup
> {
  return combineCodec(
    getMessageAddressTableLookupEncoder(),
    getMessageAddressTableLookupDecoder()
  );
}